backend/requirements.txt
# SonicMuse Backend Requirements

# Core Framework
fastapi==0.115.0
uvicorn==0.30.6

# PyTorch and Audio Processing
torch==2.3.1
torchaudio==2.3.1
audiocraft==1.3.0
faster-whisper==1.0.3
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.4
pydub==0.25.1
crepe==0.0.15

# Hugging Face Models
transformers==4.43.3
huggingface_hub==0.23.4

# Google Gemini (optional)
google-generativeai==0.7.2

# Environment and Utilities
python-dotenv==1.0.1

# Development
pytest==7.4.3
requests==2.31.0
